Move gang scheduing setting to common
