+++ pkg/common/util/client.go
@@ -1,0 +1,10 @@
+package util
+
+import "os"
+
+func homeDir() string {
+	if h := os.Getenv("HOME"); h != "" {
+		return h
+	}
+	return os.Getenv("USERPROFILE") // windows
+}
